requirements.txt
pandas==1.5.3
Pillow==9.4.0
scikit_learn==1.2.2
streamlit==1.19.0
